Improve TS tests
