+++ crates/typescript-tests/src/custom_section.ts
@@ -1,0 +1,3 @@
+import * as wbg from '../pkg/typescript_tests';
+
+const height: wbg.Height = new wbg.Person();
