+++ crates/typescript-tests/src/memory.ts
@@ -1,0 +1,3 @@
+import * as wasm from '../pkg/typescript_tests_bg';
+
+const memory: WebAssembly.Memory = wasm.memory;
