+++ crates/typescript-tests/src/opt_args_and_ret.rs
@@ -1,0 +1,6 @@
+use wasm_bindgen::prelude::*;
+
+#[wasm_bindgen]
+pub fn opt_fn(_a: Option<i32>) -> Option<i32> {
+    None
+}
