+++ crates/typescript-tests/src/opt_args_and_ret.ts
@@ -1,0 +1,3 @@
+import * as wbg from '../pkg/typescript_tests';
+
+const opt_fn: (a: number | undefined) => number | undefined = wbg.opt_fn;
