+++ crates/typescript-tests/src/simple_fn.rs
@@ -1,0 +1,4 @@
+use wasm_bindgen::prelude::*;
+
+#[wasm_bindgen]
+pub fn greet(_: &str) {}
