+++ crates/typescript-tests/src/simple_fn.ts
@@ -1,0 +1,5 @@
+import * as wbg from '../pkg/typescript_tests';
+import * as wasm from '../pkg/typescript_tests_bg';
+
+const wbg_greet: (a: string) => void = wbg.greet;
+const wasm_greet: (a: number, b: number) => void = wasm.greet;
