+++ crates/typescript-tests/src/simple_struct.ts
@@ -1,0 +1,6 @@
+import * as wbg from '../pkg/typescript_tests';
+
+const a = new wbg.A();
+wbg.A.other();
+a.foo();
+a.free();
